requirements.txt
Flask==2.2.5
fpdf==1.7.2
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
pydicom==3.0.1
torch==2.5.1
torchvision==0.20.1
gunicorn



